+++ nn_dataflow/tests/partition_test/test_proc_data_range.py
@@ -27,11 +27,11 @@ from nn_dataflow.core import ParallelEnum as pe
 
 from . import TestPartitionFixture
 
-class TestPartLayerFmapRange(TestPartitionFixture):
-    ''' Tests for part_layer_i/ofmap_range functions. '''
+class TestProcDataRange(TestPartitionFixture):
+    ''' Tests for proc_data_range functions. '''
 
-    def test_o_count(self):
-        ''' ofmap count. '''
+    def test_io_count(self):
+        ''' i/ofmap count. '''
         for wlkey in self.layers:
             layer = self.layers[wlkey]
 
@@ -39,13 +39,31 @@ class TestPartLayerFmapRange(TestPartitionFixture):
 
                 for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey):
 
-                    cnts = Counter(partition.part_layer_ofmap_range(
-                        layer, self.batch_size, part, pidx)
-                                   for pidx in part.gen_pidx())
+                    i_frngs = []
+                    o_frngs = []
 
-                    self.assertEqual(len(cnts),
+                    for pidx in part.gen_pidx():
+                        _, ifrng, ofrng = partition.proc_data_range(
+                            layer, self.batch_size, part, pidx)
+
+                        i_frngs.append(ifrng)
+                        o_frngs.append(ofrng)
+
+                    i_cnts = Counter(i_frngs)
+                    o_cnts = Counter(o_frngs)
+
+                    if isinstance(layer, ConvLayer):
+                        pidx_per_rng = part.size(pe.OUTP)
+                    elif isinstance(layer, LocalRegionLayer):
+                        pidx_per_rng = 1
+
+                    self.assertEqual(len(i_cnts), part.size() // pidx_per_rng)
+                    for v in i_cnts.values():
+                        self.assertEqual(v, pidx_per_rng)
+
+                    self.assertEqual(len(o_cnts),
                                      part.size() // part.size(pe.INPP))
-                    for v in cnts.values():
+                    for v in o_cnts.values():
                         self.assertEqual(v, part.size(pe.INPP))
 
     def test_o_no_overlap(self):
@@ -62,7 +80,7 @@ class TestPartLayerFmapRange(TestPartitionFixture):
 
                     for pidx in part.gen_pidx():
 
-                        fr = partition.part_layer_ofmap_range(
+                        _, _, fr = partition.proc_data_range(
                             layer, self.batch_size, part, pidx)
 
                         for fr2 in fr_list:
@@ -73,9 +91,9 @@ class TestPartLayerFmapRange(TestPartitionFixture):
 
                         fr_list.append(fr)
 
-    def test_o_full_layer(self):
-        ''' ofmap full layer. '''
-        for wlkey in self.layers:
+    def test_io_full_layer(self):
+        ''' i/ofmap full layer. '''
+        for wlkey in ['SM', 'POOL']:
             layer = self.layers[wlkey]
 
             for dnkey in self.dim_nodes:
@@ -83,24 +101,38 @@ class TestPartLayerFmapRange(TestPartitionFixture):
                 for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey,
                                                 optkey='NOINPP'):
 
-                    frmap = FmapRangeMap()
+                    # Remove ifmap point from full set.
+                    ifp_set = set(FmapRange(fp_beg=(0, 0, 0, 0),
+                                            fp_end=(self.batch_size,
+                                                    layer.nifm,
+                                                    layer.hifm,
+                                                    layer.wifm)).range())
+                    # Add ofmap ranges to a map.
+                    ofrmap = FmapRangeMap()
 
                     for pidx in part.gen_pidx():
 
-                        fr = partition.part_layer_ofmap_range(
+                        _, ifrng, ofrng = partition.proc_data_range(
                             layer, self.batch_size, part, pidx)
-                        frmap.add(fr, 0)
 
-                    self.assertTrue(frmap.is_complete())
+                        for ifp in ifrng.range():
+                            ifp_set.discard(ifp)
 
-                    cfr = frmap.complete_fmap_range()
+                        ofrmap.add(ofrng, 0)
+
+                    # Ifmap point set should be empty now.
+                    self.assertFalse(ifp_set)
+
+                    # Ofmap range map should be full now.
+                    self.assertTrue(ofrmap.is_complete())
+                    cfr = ofrmap.complete_fmap_range()
                     self.assertEqual(cfr.size('b'), self.batch_size)
                     self.assertEqual(cfr.size('n'), layer.nofm)
                     self.assertEqual(cfr.size('h'), layer.hofm)
                     self.assertEqual(cfr.size('w'), layer.wofm)
 
-    def test_o_equal_size(self):
-        ''' ofmap approximately equal range size. '''
+    def test_io_equal_size(self):
+        ''' i/ofmap approximately equal range size. '''
         for wlkey in self.layers:
             layer = self.layers[wlkey]
 
@@ -109,91 +141,20 @@ class TestPartLayerFmapRange(TestPartitionFixture):
                 for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey,
                                                 optkey='NOINPP'):
 
-                    size_list = []
+                    ifr_list = []
+                    osz_list = []
 
                     for pidx in part.gen_pidx():
 
-                        fr = partition.part_layer_ofmap_range(
+                        _, ifrng, ofrng = partition.proc_data_range(
                             layer, self.batch_size, part, pidx)
 
-                        size_list.append([fr.size(d) for d in 'bnhw'])
+                        ifr_list.append(ifrng)
+                        osz_list.append([ofrng.size(d) for d in 'bnhw'])
 
-                    for dsz_list, d in zip(zip(*size_list), 'bnhw'):
-                        self.assertEqual(len(dsz_list), part.size())
-                        self.assertLessEqual(max(dsz_list) - min(dsz_list), 1,
-                                             'test_o_equal_size: {} {}: '
-                                             'dim {} range size diverges. '
-                                             'max {} min {}'
-                                             .format(wlkey, dnkey, d,
-                                                     max(dsz_list),
-                                                     min(dsz_list)))
-
-    def test_i_count(self):
-        ''' ifmap count. '''
-        for wlkey in self.layers:
-            layer = self.layers[wlkey]
-
-            for dnkey in self.dim_nodes:
-
-                for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey):
-
-                    cnts = Counter(partition.part_layer_ifmap_range(
-                        layer, self.batch_size, part, pidx)
-                                   for pidx in part.gen_pidx())
-
-                    if isinstance(layer, ConvLayer):
-                        pidx_per_rng = part.size(pe.OUTP)
-                    elif isinstance(layer, LocalRegionLayer):
-                        pidx_per_rng = 1
-
-                    self.assertEqual(len(cnts), part.size() // pidx_per_rng)
-                    for v in cnts.values():
-                        self.assertEqual(v, pidx_per_rng)
-
-    def test_i_full_layer(self):
-        ''' ifmap full layer. '''
-        for wlkey in ['BASE', 'POOL']:
-            layer = self.layers[wlkey]
-
-            for part in self._gen_partition(wlkey=wlkey, optkey='NOINPP'):
-
-                # Remove point from full set.
-                fp_set = set(FmapRange(fp_beg=(0, 0, 0, 0),
-                                       fp_end=(self.batch_size,
-                                               layer.nifm,
-                                               layer.hifm,
-                                               layer.wifm)).range())
-
-                for pidx in part.gen_pidx():
-
-                    fr = partition.part_layer_ifmap_range(
-                        layer, self.batch_size, part, pidx)
-
-                    for fp in fr.range():
-                        fp_set.discard(fp)
-
-                self.assertFalse(fp_set)
-
-    def test_i_equal_size(self):
-        ''' ifmap approximately equal range size. '''
-        for wlkey in self.layers:
-            layer = self.layers[wlkey]
-
-            for dnkey in self.dim_nodes:
-
-                for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey,
-                                                optkey='NOINPP'):
-
-                    fr_list = []
-
-                    for pidx in part.gen_pidx():
-
-                        fr = partition.part_layer_ifmap_range(
-                            layer, self.batch_size, part, pidx)
-                        fr_list.append(fr)
-
+                    # For ifmaps.
                     for d in 'bnhw':
-                        dsz_list = [fr.size(d) for fr in fr_list]
+                        dsz_list = [ifrng.size(d) for ifrng in ifr_list]
 
                         thr = 1
                         if isinstance(layer, LocalRegionLayer):
@@ -213,6 +174,17 @@ class TestPartLayerFmapRange(TestPartitionFixture):
                                                      max(dsz_list),
                                                      min(dsz_list)))
 
+                    # For ofmaps.
+                    for dsz_list, d in zip(zip(*osz_list), 'bnhw'):
+                        self.assertEqual(len(dsz_list), part.size())
+                        self.assertLessEqual(max(dsz_list) - min(dsz_list), 1,
+                                             'test_o_equal_size: {} {}: '
+                                             'dim {} range size diverges. '
+                                             'max {} min {}'
+                                             .format(wlkey, dnkey, d,
+                                                     max(dsz_list),
+                                                     min(dsz_list)))
+
     def test_match_io_fmap_range(self):
         ''' ofmap and ifmap range match. '''
         for wlkey in self.layers:
@@ -225,32 +197,31 @@ class TestPartLayerFmapRange(TestPartitionFixture):
 
                     for pidx in part.gen_pidx():
 
-                        ofr = partition.part_layer_ofmap_range(
-                            layer, self.batch_size, part, pidx)
-                        ifr = partition.part_layer_ifmap_range(
+                        _, ifrng, ofrng = partition.proc_data_range(
                             layer, self.batch_size, part, pidx)
 
-                        self.assertEqual(ofr.size('b'), ifr.size('b'))
+                        self.assertEqual(ofrng.size('b'), ifrng.size('b'))
 
                         if isinstance(layer, ConvLayer):
-                            ol = ConvLayer(nifm=ifr.size('n'),
-                                           nofm=ofr.size('n'),
-                                           sofm=(ofr.size('h'), ofr.size('w')),
+                            ol = ConvLayer(nifm=ifrng.size('n'),
+                                           nofm=ofrng.size('n'),
+                                           sofm=(ofrng.size('h'),
+                                                 ofrng.size('w')),
                                            sfil=(layer.hfil, layer.wfil),
                                            strd=(layer.htrd, layer.wtrd))
                             il = ol.input_layer()
-                            self.assertEqual(il.nofm, ifr.size('n'))
+                            self.assertEqual(il.nofm, ifrng.size('n'))
                         elif isinstance(layer, LocalRegionLayer):
-                            nofm_beg, nofm_end = ofr.beg_end('n')[0]
-                            nifm_beg, nifm_end = ifr.beg_end('n')[0]
+                            nofm_beg, nofm_end = ofrng.beg_end('n')
+                            nifm_beg, nifm_end = ifrng.beg_end('n')
                             self.assertEqual(nifm_beg, max(0, \
                                     nofm_beg - layer.nreg // 2))
                             self.assertEqual(nifm_end, min(layer.nifm, \
                                     nofm_end + layer.nreg - layer.nreg // 2))
 
-                            ol = LocalRegionLayer(nofm=ofr.size('n'),
-                                                  sofm=(ofr.size('h'),
-                                                        ofr.size('w')),
+                            ol = LocalRegionLayer(nofm=ofrng.size('n'),
+                                                  sofm=(ofrng.size('h'),
+                                                        ofrng.size('w')),
                                                   nreg=layer.nreg,
                                                   sreg=(layer.hreg,
                                                         layer.wreg),
@@ -258,6 +229,27 @@ class TestPartLayerFmapRange(TestPartitionFixture):
                                                         layer.wtrd))
                             il = ol.input_layer()
 
-                        self.assertEqual(il.hofm, ifr.size('h'))
-                        self.assertEqual(il.wofm, ifr.size('w'))
+                        self.assertEqual(il.hofm, ifrng.size('h'))
+                        self.assertEqual(il.wofm, ifrng.size('w'))
+
+    def test_filrng(self):
+        ''' Filter ranges. '''
+        for wlkey in self.layers:
+            layer = self.layers[wlkey]
+
+            for dnkey in self.dim_nodes:
+
+                for part in self._gen_partition(wlkey=wlkey, dnkey=dnkey):
+
+                    for pidx in part.gen_pidx():
+                        filrng, ifrng, ofrng = partition.proc_data_range(
+                            layer, self.batch_size, part, pidx)
+
+                        self.assertEqual(len(filrng), 2)
+                        if isinstance(layer, ConvLayer):
+                            self.assertEqual(filrng[0].size(), ifrng.size('n'))
+                            self.assertEqual(filrng[1].size(), ofrng.size('n'))
+                        elif isinstance(layer, LocalRegionLayer):
+                            self.assertTrue(filrng[0].empty())
+                            self.assertTrue(filrng[1].empty())