+++ nn_dataflow/tests/partition_test/test_unit_nhops_to_proc_region.py
@@ -18,23 +18,20 @@ You should have received a copy of the Modified BSD-3 License along with this
 program. If not, see <https://opensource.org/licenses/BSD-3-Clause>.
 """
 
-import itertools
-
 from nn_dataflow.core import partition
 from nn_dataflow.core import ConvLayer
 from nn_dataflow.core import DataCategoryEnum as de
 from nn_dataflow.core import DataLayout
-from nn_dataflow.core import FmapRange, FmapRangeMap
+from nn_dataflow.core import FmapRange
 from nn_dataflow.core import NodeRegion
 from nn_dataflow.core import ParallelEnum as pe
 from nn_dataflow.core import PartitionScheme
 from nn_dataflow.core import PhyDim2
-from nn_dataflow import util
 
 from . import TestPartitionFixture
 
-class TestPartLayerUnitNhops(TestPartitionFixture):
-    ''' Tests for part_layer_unit_nhops function. '''
+class TestUnitNhopsToProcRegion(TestPartitionFixture):
+    ''' Tests for unit_nhops_to_proc_region function. '''
 
     def test_small(self):
         ''' Small case with hand calculation. '''
@@ -71,21 +68,21 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
         nr = NodeRegion(origin=PhyDim2(0, 0), dim=part.dim(),
                         type=NodeRegion.PROC)
 
-        # (0, 0, 0, 0): (0, 0, 0, 0) -- (4, 6, 18, 9): (-2, -2)
-        # (0, 0, 0, 1): (0, 0, 0, 9) -- (4, 6, 18, 18): (-2, -1)
-        # (1, 0, 0, 0): (4, 0, 0, 0) -- (8, 6, 18, 9): (-1, -2)
-        # (1, 0, 0, 1): (4, 0, 0, 9) -- (8, 6, 18, 18): (-1, -1)
+        # (0, 0, 0, 0) -- (4, 6, 18, 9): (-2, -2)
+        # (0, 0, 0, 9) -- (4, 6, 18, 18): (-2, -1)
+        # (4, 0, 0, 0) -- (8, 6, 18, 9): (-1, -2)
+        # (4, 0, 0, 9) -- (8, 6, 18, 18): (-1, -1)
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-2, -2),
-            (2, 1, 1, 2), PhyDim2(2, 2))
+            (2, 1), (1, 1), PhyDim2(2, 2))
 
-        # (0, 0, 0, 0): (0, 0, 0, 0) -- (8, 4, 8, 16): (2, 2)
-        # (0, 0, 1, 0): (0, 0, 8, 0) -- (8, 4, 16, 16): (2, 3)
-        # (0, 1, 0, 0): (0, 4, 0, 0) -- (8, 8, 8, 16): (3, 2)
-        # (0, 1, 1, 0): (0, 4, 8, 0) -- (8, 8, 16, 16): (3, 3)
+        # (0, 0, 0, 0) -- (8, 4, 16, 8): (2, 2)
+        # (0, 0, 0, 8) -- (8, 4, 16, 16): (2, 3)
+        # (0, 4, 0, 0) -- (8, 8, 16, 8): (3, 2)
+        # (0, 4, 0, 8) -- (8, 8, 16, 16): (3, 3)
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(2, 2),
-            (1, 2, 2, 1), PhyDim2(2, 2))
+            (1, 1), (2, 1), PhyDim2(2, 2))
 
         filter_nodes = [PhyDim2(0, 0)]
 
@@ -109,15 +106,15 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
         #                 -> (1, 3), (3, 3): (8 - 4) * 6 * (18 - 8) * (18 - 9)
 
         # ofmap: (2, 2) -> (0, 0):
-        #               -> (0, 1):
         #               -> (0, 2):
-        #               -> (0, 3): 4 * 4 * 8 * 8
-        #        (2, 3) -> (1, 0/1/2/3)
-        #        (3, 2) -> (2, 0/1/2/3)
-        #        (3, 3) -> (3, 0/1/2/3)
+        #               -> (1, 0):
+        #               -> (1, 2): 4 * 4 * 8 * 8
+        #        (2, 3) -> (0/1, 1/3)
+        #        (3, 2) -> (2/3, 0/2)
+        #        (3, 3) -> (2/3, 1/3)
 
-        nhops = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         self.assertEqual(nhops[de.FIL],
@@ -133,8 +130,8 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
                                        + (4 + 6) * 1 + (5 + 7) * 9
                                        + (5 + 7) * 1 + (6 + 8) * 9))
         self.assertEqual(nhops[de.OFM],
-                         4 * 4 * 8 * 8 * ((4 + 3 + 2 + 3) + (4 + 3 + 2 + 1)
-                                          + (3 + 2 + 1 + 2) + (3 + 2 + 1 + 0)))
+                         4 * 4 * 8 * 8 * ((4 + 2 + 3 + 1) + (4 + 2 + 3 + 1)
+                                          + (3 + 1 + 2 + 0) + (3 + 1 + 2 + 0)))
 
     def test_conv_layer(self):
         ''' CONV layers. '''
@@ -148,20 +145,20 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-3, -3),
-            (2, 4, 2, 2), PhyDim2(8, 4))
+            (1, 2), (4, 1), PhyDim2(8, 4))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(5, 5),
-            (1, 2, 2, 2), PhyDim2(2, 4))
+            (1, 1), (1, 2), PhyDim2(2, 4))
 
         filter_nodes = [PhyDim2(0, 0), PhyDim2(7, 7)]
 
-        nhops = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         true_nhops = self._true_unit_nhops(
-            layer, part, nr, filter_nodes, ilayout, olayout)
+            layer, nr, part, filter_nodes, ilayout, olayout)
 
         self.assertListEqual(nhops, true_nhops)
 
@@ -177,20 +174,20 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-3, 10),
-            (2, 4, 1, 2), PhyDim2(4, 4))
+            (1, 2), (4, 1), PhyDim2(4, 4))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(1, 1),
-            (2, 2, 1, 1), PhyDim2(2, 2))
+            (2, 1), (1, 2), PhyDim2(2, 2))
 
         filter_nodes = [PhyDim2(0, 0), PhyDim2(0, 7)]
 
-        nhops = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         true_nhops = self._true_unit_nhops(
-            layer, part, nr, filter_nodes, ilayout, olayout)
+            layer, nr, part, filter_nodes, ilayout, olayout)
 
         self.assertListEqual(nhops, true_nhops)
 
@@ -206,20 +203,20 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-3, -3),
-            (2, 4, 2, 2), PhyDim2(8, 4))
+            (1, 2), (4, 1), PhyDim2(8, 4))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(5, 5),
-            (1, 2, 2, 2), PhyDim2(2, 4))
+            (1, 2), (1, 1), PhyDim2(2, 4))
 
         filter_nodes = []
 
-        nhops = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         true_nhops = self._true_unit_nhops(
-            layer, part, nr, filter_nodes, ilayout, olayout)
+            layer, nr, part, filter_nodes, ilayout, olayout)
 
         self.assertListEqual(nhops, true_nhops)
 
@@ -235,20 +232,20 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-3, -3),
-            (2, 4, 2, 2), PhyDim2(8, 4))
+            (1, 2), (4, 1), PhyDim2(8, 4))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(5, 5),
-            (1, 2, 2, 2), PhyDim2(2, 4))
+            (1, 1), (1, 2), PhyDim2(2, 4))
 
         filter_nodes = []
 
-        nhops = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         true_nhops = self._true_unit_nhops(
-            layer, part, nr, filter_nodes, ilayout, olayout)
+            layer, nr, part, filter_nodes, ilayout, olayout)
 
         self.assertListEqual(nhops, true_nhops)
 
@@ -264,16 +261,16 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-3, -3),
-            (1, 1, 1, 1), PhyDim2(1, 1))
+            (1, 1), (1, 1), PhyDim2(1, 1))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(3, 3),
-            (1, 1, 1, 1), PhyDim2(1, 1))
+            (1, 1), (1, 1), PhyDim2(1, 1))
 
         filter_nodes = [PhyDim2(3, -3)]
 
-        nhops_1 = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops_1 = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         nr = NodeRegion(origin=PhyDim2(6, 6), dim=part.dim(),
@@ -281,85 +278,79 @@ class TestPartLayerUnitNhops(TestPartitionFixture):
 
         ilayout = self._make_data_layout(
             layer.nifm, layer.hifm, layer.wifm, PhyDim2(-6, -6),
-            (1, 1, 1, 1), PhyDim2(1, 1))
+            (1, 1), (1, 1), PhyDim2(1, 1))
 
         olayout = self._make_data_layout(
             layer.nofm, layer.hofm, layer.wofm, PhyDim2(6, 6),
-            (1, 1, 1, 1), PhyDim2(1, 1))
+            (1, 1), (1, 1), PhyDim2(1, 1))
 
         filter_nodes = [PhyDim2(6, -6)]
 
-        nhops_2 = partition.part_layer_unit_nhops(
-            layer, self.batch_size, part, nr,
+        nhops_2 = partition.unit_nhops_to_proc_region(
+            layer, self.batch_size, nr, part,
             filter_nodes, ilayout, olayout, self.options['BASE'])
 
         self.assertListEqual(nhops_2, [n * 2 for n in nhops_1])
 
-    def _make_data_layout(self, nfm, hfm, wfm, origin, nums, dims):
+    def _make_data_layout(self, nfm, hfm, wfm, origin, bdim, ndim, dims):
         ''' Make a DataLayout instance. '''
-        assert util.prod(nums) == dims.size()
+        frng = FmapRange((0,) * 4, (self.batch_size, nfm, hfm, wfm))
 
-        def _coord(idxs):
-            # In the order of n, b, w, h, i.e., 1, 0, 3, 2.
-            cflat = 0
-            for i in [1, 0, 3, 2]:
-                cflat = cflat * nums[i] + idxs[i]
-            assert cflat < dims.size()
-            return PhyDim2(*divmod(cflat, dims.w))
+        region = NodeRegion(origin=origin, dim=dims, type=NodeRegion.DATA)
 
-        sizes = (self.batch_size, nfm, hfm, wfm)
+        # From top to bottom: h, w, b, n.
+        order = (pe.OFMP, pe.BATP, pe.OUTP, pe.INPP)
+        pdims = [None] * pe.NUM
 
-        frmap = FmapRangeMap()
+        pdims[pe.BATP] = PhyDim2(*bdim)
+        pdims[pe.OUTP] = PhyDim2(*ndim)
+        pdims[pe.OFMP] = PhyDim2(h=dims.h // bdim[0] // ndim[0],
+                                 w=dims.w // bdim[1] // ndim[1])
+        pdims[pe.INPP] = PhyDim2(1, 1)
 
-        for idxs in itertools.product(*[range(n) for n in nums]):
+        part = PartitionScheme(order=order, pdims=pdims)
 
-            begs = [i * s // n for i, n, s in zip(idxs, nums, sizes)]
-            ends = [(i + 1) * s // n for i, n, s in zip(idxs, nums, sizes)]
+        return DataLayout(frngs=(frng,), regions=(region,), parts=(part,))
 
-            frmap.add(FmapRange(begs, ends), (_coord(idxs),))
-
-        dl = DataLayout(frmap=frmap, origin=origin, type=NodeRegion.DATA)
-        assert dl.frmap.complete_fmap_range().size() == util.prod(sizes)
-
-        return dl
-
-    def _true_unit_nhops(self, layer, part, region, filnodes, ilayout, olayout):
+    def _true_unit_nhops(self, layer, region, part, filnodes, ilayout, olayout):
         '''
         Calculate the unit number of hops for i/ofmaps.
         '''
         nhops = [0] * de.NUM
 
+        ifrmap = ilayout.fmap_range_map()
+        ofrmap = olayout.fmap_range_map()
+
         for pidx in part.gen_pidx():
 
+            # Current node.
             coord = part.coordinate(region, pidx)
+
             # Middle node of INPP.
             midpidx = list(pidx)
             midpidx[pe.INPP] = divmod(part.size(pe.INPP) // 2,
                                       part.dim(pe.INPP).w)
             midcoord = part.coordinate(region, midpidx)
 
-            # Ifmaps.
-            ifr = partition.part_layer_ifmap_range(
+            fil, ifr, ofr = partition.proc_data_range(
                 layer, self.batch_size, part, pidx)
-            for srcs, size in ilayout.frmap.rget_counter(ifr).items():
-                nhops[de.IFM] += sum(coord.hop_dist(s + ilayout.origin)
-                                     for s in srcs) * size
+
+            # Ifmaps.
+            for src, size in ifrmap.rget_counter(ifr).items():
+                nhops[de.IFM] += coord.hop_dist(src) * size
 
             # Ofmaps.
-            ofr = partition.part_layer_ofmap_range(
-                layer, self.batch_size, part, pidx)
             if coord == midcoord:
                 # From/to sources.
-                for srcs, size in olayout.frmap.rget_counter(ofr).items():
-                    nhops[de.OFM] += sum(coord.hop_dist(s + olayout.origin)
-                                         for s in srcs) * size
+                for src, size in ofrmap.rget_counter(ofr).items():
+                    nhops[de.OFM] += coord.hop_dist(src) * size
             else:
                 # From/to middle node.
                 nhops[de.OFM] += ofr.size() * coord.hop_dist(midcoord) / 2
 
             # Filters.
             if filnodes:
-                num_fils = ifr.size('n') * ofr.size('n')
+                num_fils = fil[0].size() * fil[1].size()
                 nhops[de.FIL] += num_fils * layer.filter_size() \
                         * min(coord.hop_dist(n) for n in filnodes)